requirements.txt
django
django_rest_framework
django_nose
coverage